# src/openapi.yaml
openapi: 3.0.3
info:
  title: Banking Accounts API
  description: Accounts Service provides account management features for Banco Rocks, including account creation, balance inquiry, and customer data integration.
  version: 0.1.0

servers:
  - url: http://localhost:3003

paths:
  /v1/accounts/{accountId}:
    get:
      summary: Get account details
      operationId: getAccount
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    OpenAccountResponse:
      type: object
      required:
        - accountId
        - number
        - status
        - createdAt
        - name
      properties:
        accountId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        number:
          type: string
          pattern: "^[0-9]{6}$"
          example: "123456"
        branch:
          type: string
          pattern: "^[0-9]{4}$"
          example: "0001"
        name:
          type: string
          minLength: 3
          example: "João da Silva"
        status:
          type: string
          enum: [active]
        createdAt:
          type: string
          format: date-time
          example: "2025-08-29T12:00:00Z"

    Account:
      allOf:
        - $ref: "#/components/schemas/OpenAccountResponse"
        - type: object
          required:
            - document
          properties:
            document:
              type: string
              pattern: "^[0-9]{11}$"
              example: "00000000000"
            productCode:
              type: string
              nullable: true
              example: "CA_BASIC"
            currency:
              type: string
              nullable: true
              example: "BRL"

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string